+++ djangomaat/management/commands/populate_maat_ranking.py
@@ -1,10 +1,12 @@
 from __future__ import unicode_literals
 
-from optparse import make_option
-from string import strip
-
 from django.core.management.base import BaseCommand
-from django.db.models.loading import get_model
+try:
+    from django.apps import apps
+    get_model = apps.get_model
+except ImportError:
+    # Django < 1.9
+    from django.db.models import get_model
 
 from djangomaat.register import maat
 
@@ -12,12 +14,16 @@ from djangomaat.register import maat
 class Command(BaseCommand):
     args = '[<app_label.model_name app_label.model_name ...>]'
     help = "Delete the old ranking and store the new ones."
-    option_list = BaseCommand.option_list + (
-        make_option('-s', '--simulate',
+
+    def add_arguments(self, parser):
+        parser.add_argument('models', nargs='*', type=str)
+        
+        parser.add_argument(
+            '--simulate',
             action='store_true',
             dest='simulate',
             default=False,
-            help='Simulation mode'),
+            help='Simulation mode'
         )
 
     def _parse(self, model):
@@ -33,7 +39,7 @@ class Command(BaseCommand):
         app_label, model_name = bits[0].split('.')
         typologies = None
         if len(bits) > 1:
-            typologies = map(strip, bits[1].split(','))
+            typologies = [bit.strip() for bit in bits[1].split(',')]
             if not typologies:
                 raise SyntaxError('Missing typologyes values after colon')
         return app_label, model_name, typologies
@@ -46,6 +52,8 @@ class Command(BaseCommand):
             logger = self.stdout
         else:
             logger = None
+
+        models = models or options.get('models')
 
         if models:
             for model in models: